Add method to retrieve list of exposures and make cutout method work for full CCD cutout.
The cfis_exposure_list method was part of the job builder script but might be generically useful so moved to
storage module.  Also, this now supports time ranges so data can be processed per qrunid.

The cutout method was previously modified to work with the image display system and to have
a more uniform cutout interface but those changes didn't work for the full extension cutout so
some patch work was needed.