Add a keyboard shortcut legend in a text area
Using the text area is the easiest way to get a legend displayed. There
exists a "Label" widget which just functions as a line of text to be
displayed, but it does not support newline characters and it would be
much more taxing to create a new label for every line needed.
The legend has been placed in the container next to the header text
area, as this cut down on some unused space due to the extension of the
boxes to the edge of the screen.

It seems that widgets are unable to be resized using their resize
methods. Adding the header box to another container shrinks the box
to an unusable size which requires manual changing every time. For that
reason it has been left to extend down the side of the viewer.

The console command box still isn't functioning to take all logging
outputs. It extends from one edge of the window to the other and is
only resizable manually. Optimally, the text are would be resized
upon creation to better display the area, as of now its height is
smaller than desired.