Add mosaic display for multi extension HDU lists
Only works to combine two imageHDU's. The result comes out spliced
correctly, but the image is black & white. Not sure if this is
happening because of the uncalibrated images, or something with
the mosaic_inline function that's being used to created the mosaic.
The aperture marking also gets thrown off, which shouldn't be too
difficult of a fix.

Also minor changes to the GUI and previous candidate set button.