Add ability to specify directories in VOSpace for getting json files
Originally the files were looked up directly in
/cfis/solar_system/catalogs. Now there is a button (placed at the
bottom of the GUI) which will accept a string which will be joined to
the original location.

More output is being written to the console text area in the GUI.
Was not able to get the stream handler working. Might have to do with
the text areas methods, as the stream handler can only send logging
output to methods with "write() and flush() methods" (as per the
documentation).

Submitted (and merged) a pull request to Ginga which dealt with the
faulty resize method for the text entry areas. Upon the next release,
the text entry boxes will be shortened down to a better size. Their
current lengths are what drove the placement of the newer text box
to be placed at the bottom. Once resizeable, both boxes can be placed
side by side.