Add a comparison image to the stack of images to blink if offset between images is large.
Sometimes the offset between images of a source are large and blinking between them
doesn't really help explore the reality of a detection.  A comparison image is added to the
sequence for these cases, there is a bit of manipulation of which image to display
next and like that can be improved.