Check candidate set directories for completeness
For a candidate file with an accompanying directory inside its QRUNID
directory on VOSpace, the subdirectory where all the .ast files is
now being checked for completeness by comparing the amount of files
in the directory to the amount of candidates in the candidate file.

Speed could be greatly improved by accessing the amount of candidates
immediately after creating the CandidateSet. The only area where that
length value seems to be stored is in a very hidden dictionary which
is not easily accessible.

Other minor changes to how buttons react upon activation.