Check if candidate set has been investigated when first being loaded
Status check is done during the next() method call instead of during
load. This allows for a fresh server query when loading each candidate
set, but doesn't query when blinking images of the same set.
This will allow for existence checking on a step-by-step basis instead
of making the assumption that all candidates following a
non-investigated one are also not investigated.