Write .ast files exclusively in their accepted or rejected folders
Prior to this change, a newly accepted set would be written in two
different areas. If accepted, now it will only write to its accepted
directory. Upon further investigation, if the candidate is rejected,
it will write to its rejected folder and remove the file in its
accepted folder.
This does not work both ways: taking a previously rejected file and
accepting it will not remove the file from its rejected area; a new file
will just be written in its accepted directory. It would be possible
to remove the file from the rejected directory, but having a rejected
file turned over will most likely never happen.