Add user prompt after rejecting a previously accepted file
Since accepted and rejected .ast files get written to different places
and the iterative existence checker only checks one area, it's
possible for a user to have an already-accepted file be presented.
In the case where they decide to reject this file, a new prompt at the
bottom of the application will be activated (and made red) to issue
a response. The user can then decide to remove the accepted file from
its directory and upload it as a rejected file, or leave it where it is
and move on with the candidate set.
This will allow for more fluid candidate checking, specially from a
secondary inspector.

Other minor performance/logic tweaks (my mind got refreshed after
spending a few days sick and away from coding).